requirements.txt
flask>=3.0
python-dotenv>=1.0
langchain>=0.2
langchain-openai>=0.1
langchain-community>=0.2
chromadb>=0.5
pypdf>=4.0
openai>=1.0
